Juqbox.jl Documentation

Installation

The following instructions assume that you have already installed Julia (currently version 1.5.3) on your system. Before installing Juqbox, we recommend that you add the following line to your .bashprofile (or corresponding file):<br> **export JULIAPROJECT="@."**<br> This environment variable tells Julia to look for Project.toml files in your current or parent directory.

Building and testing Juqbox

shell> julia<br> julia> ]<br> (@v1.5) pkg> add https://github.com/LLNL/Juqbox.jl.git<br> (@v1.5) pkg> precompile<br> (@v1.5) pkg> test Juqbox<br> ... all tests should pass ...<br>

To exit the package manager and Julia you do<br> (@v1.5) pkg> (DEL) <br> julia> exit()

Workflow

The work flow for solving a quantum optimal control problem consists of the following general steps:

Specify the problem

Optimize

Visualize the results

1. Specifying the problem

The setup phase includes specifying

The size of the state vector

The system and control Hamiltonians

The target gate transformation

Duration of the gate and number of time steps for integrating Schroedinger's equation.

For the parameterization the control functions, you need to specify

Carrier wave frequencies

Number of B-spline coefficients in each spline

The properties of the control problem are stored in a mutable struct that is populated by calling

params = Juqbox.objparams().

The next steps are:

Assign the initial parameter vector (called pcof0 in the examples below)

Set bounds for the parameter vector to be imposed during the optimization

Allocate working arrays by calling wa = Juqbox.Working_arrays()

Assign convergence criteria and other parameters for the optimizer

Build the optimization structure by calling prob = Juqbox.setup_ipopt_problem()

2. Optimization

Once you have been assigned the params and prob objects, as well as the initial parameter vector pcof0, the optimizer is invoked by

pcof = Juqbox.run_optimizer(prob, pcof0 [, jld2_filename])

3. Visualizing the results

General properties of the optimized solution such as trace infidelity and unitary accuracy can be evaluated, and a number of figures can generated by invoking

pl = Juqbox.plot_results(params, pcof)

An array of Julia plot objects is returned in pl. These objects can be visualized on the screen

display(pl[1])

where pl[1] is the first Julia plot object. The following plot objects are populated by the script:

pl[1] Evolution of the state vector population

pl[2] Control functions in the rotating frame of reference

pl[3] Population of forbidden energy levels

pl[4] Lab frame control function(s)

pl[5] Fourier transform of the lab-frame control functions (linear scale)

pl[6] Fourier transform of the lab-frame control functions (log scale)

pl[7] Coefficients of the optimized parameter vector

pl[8] Convergence of the optimization

Examples

Examples of the setup procedure can be found in the scripts in the Juqbox.jl/examples directory. The examples are invoked by, e.g.

include("cnot1-setup.jl")

The following cases are included:

rabi-setup.jl Pi-pulse (X-gate) for a qubit, i.e. a Rabi oscillator.

cnot1-setup.jl CNOT gate for a single qudit with 4 essential and 2 guard levels.

flux-setup.jl CNOT gate for single qubit with a flux-tuning control Hamiltonian.

cnot2-setup.jl CNOT gate for a pair of coupled qubits with guard levels.

cnot3-setup.jl Cross-resonance CNOT gate for a pair of qubits that are coupled by a cavity resonator.

Note: This case reads an optimized solution from file.

Exported types

The following types are exported and available by using Juqbox.

Juqbox.Working_Arrays — Type

wa = Working_Arrays(params::objparams, nCoeff::Int64)

Constructor for the mutable struct Working_Arrays containing preallocated working arrays.

Arguments

param:: objparams: Struct with problem definition

nCoeff:: Int64: Number of parameters in optimization

Juqbox.bcparams — Type

bcpar = bcparams(T, D1, Ncoupled, Nunc, omega, pcof)

General constructor of struct bcparams for setting up B-splines with carrier waves.

bcpar = bcparams(T, D1, omega, pcof)

Simplified constructor for the case when there are no uncoupled controls and Ncoupled = size(omega,1).

Arguments

T:: Float64: Duration of spline function

D1:: Int64: Number of basis functions in each segment

Ncoupled::Int64: Number of coupled controls in the simulation

Nunc::Int64: Number of uncoupled controls in the simulation

omega::Array{Float64,2}: Carrier wave frequencies

pcof:: Array{Float64, 1}: Coefficient vector. Must have D1*Nseg elements

First dimensions of the omega array:

With no uncoupled controls, Nunc=0 and size(omega,1) = Ncoupled.

With uncoupled controls, Nunc > 0 and size(omega,1) = Ncoupled + Nunc.

Second dimension of the omega array:

size(omega, 2) = Nfreq

Ordering of the pcof array:

First consider the case without uncoupled control functions, Nunc = 0: Then the pcof array then has 2*Ncoupled*Nfreq*D1 elements. Each ctrl ∈ [1,Ncoupled] and freq ∈ [1,Nfreq] corresponds to D1 elements in the pcof vector. For the case Ncoupled = 2 and Nfreq = 2, the elements are ordered according to

ctrl	freq	α_1	α_2

1	1	1:D1	D1+1:2 D1

1	2	2 D1+1: 3 D1	3 D1+1:4 D1

2	1	4 D1+1: 5 D1	5 D1+1:6 D1

2	2	6 D1+1: 7 D1	7 D1+1: 8D1

If there are uncoupled controls, Nunc > 0, the pcof array should have (2*Ncoupled + Nunc)*Nfreq*D1 elements. The last Nunc*Nfreq*D1 elements correspond to the uncoupled control functions and are ordered in a corresponding way.

External links

Spline Wavelet on Wikipedia.

Juqbox.objparams — Type

params = objparams(Ne, Ng, T, Nsteps;
                    Uinit=Uinit, 
                    Utarget=Utarget,
                    Cfreq=Cfreq, 
                    Rfreq=Rfreq, 
                    Hconst=Hconst [, 
                    Hsym_ops=Hsym_ops,
                    Hanti_ops=Hanti_ops, 
                    Hunc_ops=Hunc_ops,
                    wmatScale=wmatScale, 
                    use_sparse=use_sparse])

Constructor for the mutable struct objparams. The sizes of the arrays in the argument list are based on Ntot = prod(Ne + Ng), Ness = prod(Ne), Nosc = length(Ne) = length(Ng).

Notes: It is assumed that length(Hsym_ops) = length(Hanti_ops) =: Ncoupled. The matrices Hconst, Hsym_ops[j]and Hanti_ops[j], for j∈[1,Ncoupled], must all be of size Ntot × Ntot. The matrices Hsym_ops[j] must be symmetric and Hanti_ops[j] must be skew-symmetric. The matrices Hunc_ops[j], for j∈[1,Nunc], where Nunc = length(Hunc_ops), must also be of size Ntot × Ntot and either be symmetric or skew-symmetric.

Arguments

Ne::Array{Int64,1}: Number of essential energy levels for each subsystem

Ng::Array{Int64,1}: Number of guard energy levels for each subsystem

T::Float64: Duration of gate

Nsteps::Int64: Number of timesteps for integrating Schroedinger's equation

Uinit::Array{Float64,2}: (keyword) Matrix holding the initial conditions for the solution matrix of size Uinit[Ntot, Ness]

Utarget::Array{Complex{Float64},2}: (keyword) Matrix holding the target gate matrix of size Uinit[Ntot, Ness]

Cfreq::Array{Float64,2}: (keyword) Carrier wave frequencies of size Cfreq[Ncoupled, Nfreq]

Rfreq::Array{Float64,2}: (keyword) Rotational frequencies of size Rfreq[Nosc]

Hconst::Array{Float64,2}: (keyword) Time-independent part of the Hamiltonian matrix of size Ntot × Ntot

Hsym_ops:: Array{Array{Float64,2},1}: (keyword) Array of symmetric control Hamiltonians, each of size Ntot × Ntot

Hanti_ops:: Array{Array{Float64,2},1}: (keyword) Array of anti-symmetric control Hamiltonians, each of size Ntot × Ntot

Hunc_ops:: Array{Array{Float64,2},1}: (keyword) Array of uncoupled control Hamiltonians, each of size Ntot × Ntot

wmatScale::Float64 = 1.0: (keyword) Scaling factor for suppressing guarded energy levels

use_sparse::Bool = false: (keyword) Set to true to sparsify all Hamiltonian matrices

Juqbox.splineparams — Type

spar = splineparams(T, D1, Nseg, pcof)

Constructor for struct splineparams, which sets up the parameters for a regular B-spline function (without carrier waves).

Arguments

T:: Float64: Duration of spline function

D1:: Int64: Number of basis functions in each spline

Nseg:: Int64: Number of splines (real, imaginary, different ctrl func)

pcof:: Array{Float64, 1}: Coefficient vector. Must have D1*Nseg elements

External links

Spline Wavelet on Wikipedia.

Exported functions

The following methods (functions) are exported and available by using Juqbox.

Juqbox.assign_thresholds — Function

minCoeff, maxCoeff = assign_thresholds(params, D1, maxpar [, maxpar_unc])

Build vector of frequency independent min/max parameter constraints for each coupled and (optionally) uncoupled control function. Here, minCoeff = -maxCoeff.

Arguments

params:: objparams: Struct containing problem definition.

D1:: Int64: Number of basis functions in each segment.

maxpar::Array{Float64,1}: Maximum parameter value for each coupled control.

maxpar_unc::Array{Float64,1}=: (optional) Maximum parameter value for each uncoupled control.

Juqbox.assign_thresholds_ctrl_freq — Method

minCoeff, maxCoeff = assign_thresholds_ctrl_freq(params, D1, maxamp)

Build vector of parameter min/max constraints that can depend on the control function and carrier wave frequency, with minCoeff = -maxCoeff, assuming no uncoupled control functions.

Arguments

params:: objparams: Struct containing problem definition.

D1:: Int64: Number of basis functions in each segment.

maxamp:: Matrix{Float64}: maxamp[c,f] is the maximum parameter value for ctrl c and frequency f

Juqbox.assign_thresholds_freq — Method

minCoeff, maxCoeff = assign_thresholds_freq(maxamp, Ncoupled, Nfreq, D1)

Build vector of frequency dependent min/max parameter constraints, with minCoeff = -maxCoeff, when there are no uncoupled control functions.

Arguments

maxamp::Array{Float64,1}: Maximum parameter value for each frequency

Ncoupled::Int64: Number of coupled controls in the simulation

Nfreq::Int64: Number of carrier wave frequencies used in the controls

D1:: Int64: Number of basis functions in each control function

Juqbox.bcarrier2 — Method

f = bcarrier2(t, params, func)

Evaluate a B-spline function with carrier waves. See also the bcparams constructor.

Arguments

t::Float64: Evaluate spline at parameter t ∈ [0, param.T]

param::params: Parameters for the spline

func::Int64: Spline function index ∈ [0, param.Nseg-1]

Juqbox.bspline2 — Method

f = bspline2(t, splineparam, splinefunc)

Evaluate a B-spline function. See also the splineparams constructor.

Arguments

t::Float64: Evaluate spline at parameter t ∈ [0, param.T]

param::splineparams: Parameters for the spline

splinefunc::Int64: Spline function index ∈ [0, param.Nseg-1]

Juqbox.calculate_timestep — Function

nsteps = calculate_timestep(T, H0, Hunc_ops, maxpar [, Pmin = 40])

Estimate the number of time steps needed for an accurate simulation, when there are no coupled controls

Arguments

T:: Float64: Final simulation

H0::Array{Float64,2}: Time-independent part of the Hamiltonian matrix

Hunc_ops:: Array{Float64,2}: Array of uncoupled control Hamiltonians

max_unc:: Array{Float64,1}: Maximum parameter value for each subsystem (uncoupled)

Pmin:: Int64: Sample rate for accuracy (assuming a slowly varying Hamiltonian)

Juqbox.calculate_timestep — Function

nsteps = calculate_timestep(T, H0, Hsym_ops, Hanti_ops, maxpar [, Pmin = 40])

Estimate the number of time steps needed for the simulation, for the case without uncoupled controls.

Arguments

T:: Float64: Final simulation

H0::Array{Float64,2}: Time-independent part of the Hamiltonian matrix

Hsym_ops:: Array{Float64,2}: Array of symmetric control Hamiltonians

Hanti_ops:: Array{Float64,2}: Array of symmetric control Hamiltonians

maxpar:: Array{Float64,1}: Maximum parameter value for each subsystem

Pmin:: Int64: Number of time steps per shortest period (assuming a slowly varying Hamiltonian).

Juqbox.calculate_timestep — Function

nsteps = calculate_timestep(T, H0, Hsym_ops, Hanti_ops, Hunc_ops, 
                                          maxpar, max_flux[, Pmin = 40])

Estimate the number of time steps needed for the simulation, when there are uncoupled controls.

Arguments

T:: Float64: Final simulation

H0::Array{Float64,2}: Time-independent part of the Hamiltonian matrix

Hsym_ops:: Array{Float64,2}: Array of symmetric control Hamiltonians

Hanti_ops:: Array{Float64,2}: Array of symmetric control Hamiltonians

Hunc_ops:: Array{Float64,2}: Array of uncoupled control Hamiltonians

maxpar:: Array{Float64,1}: Maximum parameter value for each coupled control

max_flux:: Array{Float64,1}: Maximum parameter value for each uncoupled control

Pmin:: Int64: Number of time steps per shortest period (assuming a slowly varying Hamiltonian).

Juqbox.estimate_Neumann! — Function

estimate_Neumann!(tol, params, maxpar[, maxunc])

Estimate the number of terms needed by the Neumann series approach for solving the linear system during the implicit steps of the Störmer-Verlet scheme. See also neumann!

Arguments

tol:: Float64: Error tolerance in inverting implicit SV term

params:: objparams: Struct containing problem definition

maxpar:: Array{Float64,1}: Maximum parameter value for each coupled control

maxunc:: Array{Float64,1}: (optional) Maximum parameter value for each uncoupled controls

Juqbox.evalctrl — Method

pj [, qj] = evalctrl(params, pcof0, td, func)

Evaluate the control function with index func at an array of time levels td.

NOTE: the control function index func is 1-based.

NOTE: The return value(s) depend on func. For func∈[1,Ncoupled], pj, qj are returned. Otherwise, only pj is returned, corresponding to control number func.

Arguments

params:: objparams: Struct with problem definition

pcof0:: Array{Float64,1}): Vector of parameter values

td:: Array{Float64,1}): Time values control is to be evaluated

jHam:: Int64: Index of the control signal desired

Juqbox.gradbcarrier2! — Method

gradbcarrier2!(t, params, func, g) -> g

Evaluate the gradient of a control function with respect to all coefficient.

NOTE: the index of the control functions is 0-based. For a set of coupled controls, mod(func,2)=0 corresponds to ∇ pj(t) and mod(func,2) = 1 corresponds to ∇ qj(t), where j = div(func,2).

Arguments

t::Float64: Evaluate spline at parameter t ∈ [0, param.T]

params::bcparams: Parameters for the spline

func::Int64: Control function index ∈ [0, param.Nseg-1]

g::Array{Float64,1}: Preallocated array to store calculated gradient

Juqbox.gradbspline2 — Method

g = gradbspline2(t, param, splinefunc)

Evaluate the gradient of a spline function with respect to all coefficient. NOTE: the index of the spline functions are 0-based. For a set of coupled controls, mod(splinefunc,2)=0 corresponds to ∇ pj(t) and mod(splinefunc,2) = 1 corresponds to ∇ qj(t), where j = div(splinefunc,2).

Arguments

t::Float64: Evaluate spline at parameter t ∈ [0, param.T]

param::splineparams: Spline parameter object

splinefunc::Int64: Spline function index ∈ [0, param.Nseg-1]

Juqbox.identify_forbidden_levels — Function

forbiddenlev = identify_guard_levels(params[, custom = 0])

Build a Bool array indicating which energy levels are forbidden levels in the state vector. The forbidden levels in a state vector are defined as thos corresponding to the highest energy level in at least one of its subsystems.

Arguments

params:: objparams: Struct with problem definition

custom:: Int64: For nonzero value special stirap pulses case

Juqbox.identify_guard_levels — Function

guardlev = identify_guard_levels(params[, custom = 0])

Build a Bool array indicating if a given energy level is a guard level in the simulation.

Arguments

params:: objparams: Struct with problem definition

custom:: Int64: A nonzero value gives a special stirap pulses case

Juqbox.marginalize3 — Method

marg_prob = marginalize3(params, unitaryhist)

Evaluate marginalized probabilities for the case of 3 subsystems.

Arguments

param:: objparams: Struct with problem definition

unitaryhist:: Array{Complex{Float64},3}): State vector history for each timestep

Juqbox.plot_conv_hist — Function

pconv = plot_conv_hist(params [, convname:: String=""])

Plot the optimization convergence history, including history of the different terms in the objective function and the norm of the gradient.

Arguments

param:: objparams: Struct with problem definition

convname:: String: Name of plot file to be generated

Juqbox.plot_results — Method

pl = plot_results(params, pcof; [casename = "test", savefiles = false, samplerate = 32])

Create array of plot objects that can be visualized by, e.g., display(pl[1]).

Arguments

params::objparams: Object holding problem definition

pcof::Array{Float64,1}: Parameter vector

casename::String: Default: "test". String used in plot titles and in file names

savefiles::Bool: Default: false.Set to true to save plots on files with automatically generated filenames

samplerate:: Int64: Default: 32 samples per unit time (ns). Sample rate for generating plots.

Juqbox.plotspecified — Method

plt = plotspecified(us, params, guardlev, specifiedlev)

Plot the evolution of the state vector for specified levels.

Arguments

us:: Array{Complex{Float64},3}): State vector history for each timestep

param:: objparams: Struct with problem definition

us:: Array{Bool,1}): Boolean array indicating if a certain level is a guard level

'specifiedlev:: Array{Bool,1}': Boolean array indicating which levels to be plotted

Juqbox.plotunitary — Method

plt = plotunitary(us, params, guardlev)

Plot the evolution of the state vector.

Arguments

us:: Array{Complex{Float64},3}): State vector history for each timestep

param:: objparams: Struct with problem definition

guardlev:: Array{Bool,1}): Boolean array indicating if a certain level is a guard level

Juqbox.read_pcof — Method

pcof = read_pcof(refFileName)

Read the parameter vector pcof from a JLD2 formatted file

Arguments

refFileName: String holding the name of the file.

Juqbox.run_optimizer — Function

pcof = run_optimizer(prob, pcof0 [, baseName:: String=""])

Call IPOPT to optimizize the control functions.

Arguments

prob:: IpoptProblem: Struct containing Ipopt parameters callback functions

pcof0:: Array{Float64, 1}: Initial guess for the parameter values

baseName:: String: Name of file for saving the optimized parameters; extension ".jld2" is appended

Juqbox.save_pcof — Method

save_pcof(refFileName, pcof)

Save the parameter vector pcof on a JLD2 formatted file with handle pcof

Arguments

refFileName: String holding the name of the file.

pcof: Vector of floats holding the parameters.

Juqbox.setup_ipopt_problem — Function

prob = setup_ipopt_problem(params, wa, nCoeff, minCoeff, maxCoeff [, maxIter=50, 
                        lbfgsMax=10, startFromScratch=true, ipTol=1e-5, acceptTol=1e-5, acceptIter=15])

Setup structure containing callback functions and convergence criteria for optimization via IPOPT.

Arguments

params:: objparams: Struct with problem definition

wa::Working_Arrays: Struct containing preallocated working arrays

nCoeff:: Int64: Number of parameters in optimization

minCoeff:: Array{Float64, 1}: Minimum allowable value for each parameter

maxCoeff:: Array{Float64, 1}: Maximum allowable value for each parameter

maxIter:: Int64: Maximum number of iterations to be taken by optimizer

lbfgsMax:: Int64: Maximum number of past iterates for Hessian approximation by L-BFGS

startFromScratch:: Bool: Specify whether the optimization is starting from file or not

ipTol:: Float64: Desired convergence tolerance (relative)

acceptTol:: Float64: Acceptable convergence tolerance (relative)

acceptIter:: Int64: Number of acceptable iterates before triggering termination

Juqbox.setup_rotmatrices — Method

omega1[, omega2, omega3] = setup_rotmatrices(Ne, Ng, fund_freq)

Build diagonal rotation matrices based on the |0⟩to |1⟩ transition frequency in each sub-system.

Arguments

Ne::Array{Int64,1}: Number of essential energy levels for each subsystem

Ng::Array{Int64,1}: Number of guard energy levels for each subsystem

fund_freq::Array{Float64}: Transitions frequency [GHz] for each subsystem

Juqbox.traceobjgrad — Function

objf = traceobjgrad(pcof0, params, wa[, verbose = false, evaladjoint = true])

Perform a forward and/or adjoint Schrödinger solve to evaluate the objective function and/or gradient.

Arguments

pcof0::Array{Float64,1}: Array of parameter values defining the controls

param::objparams: Struct with problem definition

wa::Working_Arrays: Struct containing preallocated working arrays

verbose::Bool = false: Run simulation with additional terminal output and store state history.

evaladjoint::Bool = true: Solve the adjoint equation and calculate the gradient of the objective function.

Juqbox.wmatsetup — Method

wmat = wmatsetup(Ne, Ng)

Build the default positive semi-definite weighting matrix W to calculate the leakage into higher energy forbidden states

Arguments

Ne::Array{Int64,1}: Number of essential energy levels for each subsystem

Ng::Array{Int64,1}: Number of guard energy levels for each subsystem

Juqbox.zero_start_end! — Method

zero_start_end!(params, D1, minCoeff, maxCoeff)

Force the control functions to start and end at zero by setting zero bounds for the first two and last two parameters in each B-spline segment.

Arguments

params:: objparams: Struct containing problem definition.

D1:: Int64: Number of basis functions in each segment.

minCoeff:: Vector{Float64}: Lower parameter bounds to be modified

maxCoeff:: Vector{Float64}: Upper parameter bounds to be modified

Index

Juqbox.Working_Arrays

Juqbox.bcparams

Juqbox.objparams

Juqbox.splineparams

Juqbox.assign_thresholds

Juqbox.assign_thresholds_ctrl_freq

Juqbox.assign_thresholds_freq

Juqbox.bcarrier2

Juqbox.bspline2

Juqbox.calculate_timestep

Juqbox.calculate_timestep

Juqbox.calculate_timestep

Juqbox.estimate_Neumann!

Juqbox.evalctrl

Juqbox.gradbcarrier2!

Juqbox.gradbspline2

Juqbox.identify_forbidden_levels

Juqbox.identify_guard_levels

Juqbox.marginalize3

Juqbox.plot_conv_hist

Juqbox.plot_results

Juqbox.plotspecified

Juqbox.plotunitary

Juqbox.read_pcof

Juqbox.run_optimizer

Juqbox.save_pcof

Juqbox.setup_ipopt_problem

Juqbox.setup_rotmatrices

Juqbox.traceobjgrad

Juqbox.wmatsetup

Juqbox.zero_start_end!